+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 - Fixed mesh generation methods.
 - Fixed Mix Data node for zero length lists.
 - Fixed conflicts with the Node Wrangler add-on keymap.
+- Fixed incorrect file loading of Shape Key sockets.
 
 ### Changed